Remove unnecessary toUShort()

+++ mesh-core/src/main/java/no/nordicsemi/kotlin/mesh/core/model/Element.kt
@@ -44,7 +44,7 @@ data class Element internal constructor(
 
     @Transient
     var unicastAddress =
-        parentNode?.primaryUnicastAddress ?: UnicastAddress(address = models.size.toUShort())
+        parentNode?.primaryUnicastAddress ?: UnicastAddress(address = models.size)
         internal set
 
     init {
